# Omit invalid ctor option