--- a/geom/ctor_options.go
+++ b/geom/ctor_options.go
@@ -20,8 +20,34 @@ func DisableAllValidations(o *ctorOptionSet) {
 	o.skipValidations = true
 }
 
+// OmitInvalid causes geometry constructors to omit any geometries or
+// sub-geometries that are invalid.
+//
+// Behaviour for each geometry type is:
+//
+// * Point and MultiPoint: no effect.
+//
+// * LineString: if the LineString is invalid (e.g. doesn't contain at least 2
+// distinct points), then it is replaced with an empty LineString.
+//
+// * MultiLineString: if a child LineString is invalid, then it is omitted from
+// the MultiLineString.
+//
+// * Polygon: if the Polygon is invalid (e.g. self intersecting rings or rings
+// that intersect in an invalid way), then it is replaced with an empty
+// Polygon.
+//
+// * MultiPolygon: if a child Polygon is invalid, then it is omitted from the
+// MultiPolygon. If two child Polygons (that weren't previously omitted)
+// interact in an invalid way, then the MultiPolygon is replaced with an empty
+// MultiPolygon.
+func OmitInvalid(o *ctorOptionSet) {
+	o.omitInvalid = true
+}
+
 type ctorOptionSet struct {
 	skipValidations bool
+	omitInvalid     bool
 }
 
 func newOptionSet(opts []ConstructorOption) ctorOptionSet {
@@ -41,4 +67,9 @@ func newOptionSet(opts []ConstructorOption) ctorOptionSet {
 func skipValidations(opts []ConstructorOption) bool {
 	os := newOptionSet(opts)
 	return os.skipValidations
+}
+
+func omitInvalid(opts []ConstructorOption) bool {
+	os := newOptionSet(opts)
+	return os.omitInvalid
 }